Extend performance diagnosing utility support to all commands
